UF-675 [FIX] Engagement lines deletion when a PO is cancelled.

+++ analytic_distribution/account_commitment.py
@@ -286,12 +286,15 @@ class account_commitment(osv.osv):
         for c in self.browse(cr, uid, ids, context=context):
             # Search analytic lines that have commitment line ids
             search_ids = self.pool.get('account.analytic.line').search(cr, uid, [('commitment_line_id', 'in', [x.id for x in c.line_ids])], context=context)
+            if not search_ids:
+                return True
             # Delete them
-            res = self.pool.get('account.analytic.line').unlink(cr,  uid, search_ids, context=context)
+            res = self.pool.get('account.analytic.line').unlink(cr, uid, search_ids, context=context)
             # And finally update commitment voucher state and lines amount
-            if res:
-                self.pool.get('account.commitment.line').write(cr, uid, [x.id for x in c.line_ids], {'amount': 0}, context=context)
-                self.write(cr, uid, [c.id], {'state':'done'}, context=context)
+            if not res:
+                raise osv.except_osv(_('Error'), _('An error occured on engagement lines deletion.'))
+            self.pool.get('account.commitment.line').write(cr, uid, [x.id for x in c.line_ids], {'amount': 0}, context=context)
+            self.write(cr, uid, [c.id], {'state':'done'}, context=context)
         return True
 
 account_commitment()
